Serve client email from API
