+++ apps/eaas-api/src/app/controllers/sarcophagi.controller.ts
@@ -10,6 +10,15 @@ const getClientSarcophagi = async (req: RequestWithUser, res: Response) => {
   res.status(200).send(clientSarcophagi);
 };
 
+const getSarcoClientEmail = async (req: RequestWithUser, res: Response) => {
+  const { sarcoId } = req.params;
+
+  const clientEmail = await sarcophagiService.getSarcoClientEmail(sarcoId);
+
+  res.status(200).send(clientEmail);
+};
+
 export const sarcophagiController = {
   getClientSarcophagi,
+  getSarcoClientEmail,
 };
