+++ apps/webapp/eaas-client/src/api/sarcophagi.ts
@@ -10,3 +10,12 @@ export async function getClientSarcophagi(): Promise<SarcophagusData[]> {
     throw handleApiError(error);
   }
 }
+
+export async function getSarcoClientEmail(sarcoId: string): Promise<string> {
+  try {
+    const response = await axios.get(`/api/sarcophagi/${sarcoId}/client-email`);
+    return response.data as string;
+  } catch (error) {
+    throw handleApiError(error);
+  }
+}
