+++ apps/webapp/eaas-client/src/ui/notFound.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Center, Text, Flex, Link, Button } from "@chakra-ui/react";
 import { useNavigate } from "react-router-dom";
 import { RouteKey, RoutesPathMap } from "./routerConstants";
